requirements.txt
aiohttp
discord.py
pycryptodome
python-dotenv
rich
tensorflow
